// src/app/model/UserBO.ts
export class UserDetailsBO{
    id: string;

    mobile : number;
    state: string;
    city: string;
    zip: number;
    occupation: string;
    joiningDate: String;
    fatherName: string;
    motherName: string;
    address: string;
    securityAmount:number;
    company: string;
    occupancy:string;
    name: string;
    roomNo: number;
    gender: string;
}